requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.3
sentence-transformers==4.1.0
faiss-cpu==1.10.0
numpy==2.0.2
openai==1.75.0
python-multipart==0.0.20
PyPDF2==3.0.1
python-docx==1.1.2
pandas==2.2.3
beautifulsoup4==4.13.4
markdown==3.8
uvicorn==0.34.2
